read/write return ssize_t

+++ src/stream.h
@@ -8,10 +8,10 @@ BEGIN_C_DECLS
 struct _stream_t;
 typedef struct _stream_t stream_t;
 
-typedef int    (*stream_read_t)(stream_t* s, void* buff, size_t len);
-typedef int    (*stream_write_t)(stream_t* s, void* buff, size_t len);
-typedef void   (*stream_flush_t)(stream_t* s);
-typedef void   (*stream_close_t)(stream_t* s);
+typedef ssize_t (*stream_read_t)(stream_t* s, void* buff, size_t len);
+typedef ssize_t (*stream_write_t)(stream_t* s, void* buff, size_t len);
+typedef void    (*stream_flush_t)(stream_t* s);
+typedef void    (*stream_close_t)(stream_t* s);
 
 struct _stream_t {
 	stream_read_t  read;
@@ -20,12 +20,12 @@ struct _stream_t {
 	stream_close_t close;
 };
 
-static inline int stream_read(stream_t* s, void* buff, size_t len) {
+static inline ssize_t stream_read(stream_t* s, void* buff, size_t len) {
 	return_value_if_fail(s != NULL && s->read != NULL && buff != NULL, 0);
 	return s->read(s, buff, len);
 }
 
-static inline int stream_write(stream_t* s, void* buff, size_t len) {
+static inline ssize_t stream_write(stream_t* s, void* buff, size_t len) {
 	return_value_if_fail(s != NULL && s->write != NULL && buff != NULL, 0);
 	return s->write(s, buff, len);
 }
